test(hooks): split the contract-surface tests out of import selection
Adding the two missing behavior tests took the module to 547 lines and the
taste count ratchet rejected the push at +1 over baseline. Caught locally in 23
seconds, before the remote matrix.

Split rather than suppressed. The ADR-104 exception earlier in this branch was
right because the linter's remediation is code-shaped and a decision record has
no helpers, types, or constants to extract; a test module has all three and the
advice applies. Suppressing here would have been the lazy read of the earlier
precedent.

The seam is real rather than convenient. The remaining module tests what the
selector and the stand-in do; the new one tests what the repository says they
do, in the three places it says it. They fail for different reasons and are read
by different people: a reader chasing why a push collected instead of executing
wants the first, a reader who changed the set of probed classes wants the
second. Nothing in the new module executes pytest, and its docstring says so,
because three surfaces agreeing is not the same property as any of them being
true.

35 tests before, 35 after.

Refs #5315
